requirements.txt
numpy
pandas
matplotlib
seaborn
torch
torchvision
tqdm
pyyaml
